enable breadcrumbs and context-scoped calls for node
