+++ test/node/features/express.feature
@@ -165,3 +165,22 @@ Scenario: adding body to request metadata
   And the "file" of stack frame 0 equals "scenarios/app.js"
   And the event "request.body.data" equals "in_request_body"
   And the event "request.httpMethod" equals "POST"
+
+Scenario: Breadcrumbs from one request do not appear in another
+  When I open the URL "http://express/breadcrumbs_a"
+  And I wait to receive an error
+  Then the error is valid for the error reporting API version "4" for the "Bugsnag Node" notifier
+  And the event has a "manual" breadcrumb named "For the first URL"
+  And the event "request.url" equals "http://express/breadcrumbs_a"
+  And the event "request.httpMethod" equals "GET"
+  And the event "request.clientIp" is not null
+  And I discard the oldest error
+
+  And I open the URL "http://express/breadcrumbs_b"
+  And I wait to receive an error
+  Then the error is valid for the error reporting API version "4" for the "Bugsnag Node" notifier
+  And the event has a "manual" breadcrumb named "For the second URL"
+  And the event does not have a "manual" breadcrumb with message "For the first URL"
+  And the event "request.url" equals "http://express/breadcrumbs_b"
+  And the event "request.httpMethod" equals "GET"
+  And the event "request.clientIp" is not null
