+++ test/node/features/fixtures/contextualize/Dockerfile
@@ -1,0 +1,11 @@
+ARG NODE_VERSION=8
+FROM node:$NODE_VERSION-alpine
+
+WORKDIR /app
+
+COPY package* ./
+RUN npm install
+
+COPY . ./
+
+RUN npm install --no-package-lock --no-save bugsnag-node*.tgz
