+++ test/node/features/fixtures/contextualize/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "bugsnag-test"
+}
