+++ test/node/features/unhandled_errors.feature
@@ -55,50 +55,6 @@ Scenario: not reporting unhandledRejections when autoDetectErrors is off
   And I wait for 1 second
   Then I should receive no requests
 
-Scenario: using contextualize to add context to an error
-  And I run the service "unhandled" with the command "node scenarios/contextualize"
-  And I wait to receive an error
-  Then the error is valid for the error reporting API version "4" for the "Bugsnag Node" notifier
-  And the event "unhandled" is true
-  And the event "severity" equals "error"
-  And the event "severityReason.type" equals "unhandledException"
-  And the exception "errorClass" equals "Error"
-  And the exception "message" equals "ENOENT: no such file or directory, open 'does not exist'"
-  And the exception "type" equals "nodejs"
-  And the "file" of stack frame 0 equals "scenarios/contextualize.js"
-  And the "lineNumber" of stack frame 0 equals 12
-  And the event "metaData.subsystem.name" equals "fs reader"
-  And the event "metaData.subsystem.widgetsAdded" equals "cat,dog,mouse"
-
-@skip_before_node_16
-Scenario: using contextualize with an unhandled rejection (with context added)
-  And I run the service "unhandled" with the command "node scenarios/contextualize-unhandled-rejection"
-  And I wait to receive an error
-  Then the error is valid for the error reporting API version "4" for the "Bugsnag Node" notifier
-  And the event "unhandled" is true
-  And the event "severity" equals "error"
-  And the event "severityReason.type" equals "unhandledPromiseRejection"
-  And the exception "errorClass" equals "Error"
-  And the exception "message" equals "unhandled rejection"
-  And the exception "type" equals "nodejs"
-  And the "file" of stack frame 0 equals "scenarios/contextualize-unhandled-rejection.js"
-  And the "lineNumber" of stack frame 0 equals 12
-  And the event "metaData.subsystem.name" equals "fs reader"
-  And the event "metaData.subsystem.widgetsAdded" equals "cat,dog,mouse"
-
-Scenario: using contextualize with an unhandled rejection (no context added)
-  And I run the service "unhandled" with the command "node scenarios/contextualize-unhandled-rejection"
-  And I wait to receive an error
-  Then the error is valid for the error reporting API version "4" for the "Bugsnag Node" notifier
-  And the event "unhandled" is true
-  And the event "severity" equals "error"
-  And the event "severityReason.type" equals "unhandledPromiseRejection"
-  And the exception "errorClass" equals "Error"
-  And the exception "message" equals "unhandled rejection"
-  And the exception "type" equals "nodejs"
-  And the "file" of stack frame 0 equals "scenarios/contextualize-unhandled-rejection.js"
-  And the "lineNumber" of stack frame 0 equals 12
-
 Scenario: overridden handled state in a callback
   And I run the service "unhandled" with the command "node scenarios/modify-unhandled-callback"
   And I wait to receive an error
